[SC-165] generalization of json editing component and using it to edit activity
